adjust data type and the pipeline data transformatioin

+++ housing_scrapy拷貝/housing_scrapy/spiders/scrapy_house.py
@@ -52,11 +52,11 @@ class ScrapyHouseSpider(scrapy.Spider):
                 item['simple_address'] = item_json.get('simple_address')
                 item['current_sale_num'] = int(item_json.get('sale_num', 0))
                 item['building_purpose'] = item_json.get('build_purpose_simple')
-                item['browse_num'] = int(item_json.get('browse_num'))
+                item['browse_num'] = item_json.get('browse_num')
                 item['rent_num'] = item_json.get('rent_num')
                 item['agent_company'] = agent_company
-                item['total_sold'] = item_json.get('price_num')
-                item['price'] = float(item_json['price']['price'])
+                item['total_sold'] = int(item_json.get('price_num'))
+                item['price'] = item_json['price']['price']
                 item['station_name'] = item_json.get('station_name', 'No station name')
                 item['latitude'] = float(item_json.get('lat'))
                 item['longitude'] = float(item_json.get('lng'))
